refactored PomodoroSession.cs and dto added

+++ backend/Controllers/PomodoroController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using StudyVerseBackend.Entities;
 using StudyVerseBackend.Services;
 
@@ -7,6 +9,7 @@ namespace StudyVerseBackend.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class PomodoroSessionController : ControllerBase
     {
         private readonly PomodoroSessionService _pomodoroSessionService;
@@ -20,6 +23,8 @@ namespace StudyVerseBackend.Controllers
         [HttpGet]
         public async Task<ActionResult<List<PomodoroSession>>> GetAllPomodoroSessions()
         {
+            var userId = GetUserIdFromToken();
+            if (userId == null) return Unauthorized("Invalid User Token");
             var sessions = await _pomodoroSessionService.GetAllPomodoroSessions();
             return Ok(sessions);
         }
@@ -28,11 +33,14 @@ namespace StudyVerseBackend.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<PomodoroSession>> GetPomodoroSession(int id)
         {
+            var userId = GetUserIdFromToken();
+            if (userId == null) return Unauthorized("Invalid User Token");
             var session = await _pomodoroSessionService.GetPomodoroSessionById(id);
             if (session == null)
             {
                 return NotFound();
             }
+
             return Ok(session);
         }
 
@@ -40,6 +48,8 @@ namespace StudyVerseBackend.Controllers
         [HttpPost]
         public async Task<ActionResult<PomodoroSession>> CreatePomodoroSession(PomodoroSession session)
         {
+            var userId = GetUserIdFromToken();
+            if (userId == null) return Unauthorized("Invalid User Token");
             var createdSession = await _pomodoroSessionService.CreatePomodoroSession(session);
             return CreatedAtAction(nameof(GetPomodoroSession), new { id = createdSession.SessionId }, createdSession);
         }
@@ -48,24 +58,36 @@ namespace StudyVerseBackend.Controllers
         [HttpPut("{id}")]
         public async Task<ActionResult<PomodoroSession>> UpdatePomodoroSession(int id, PomodoroSession updatedSession)
         {
+            var userId = GetUserIdFromToken();
+            if (userId == null) return Unauthorized("Invalid User Token");
             var session = await _pomodoroSessionService.UpdatePomodoroSession(id, updatedSession);
             if (session == null)
             {
                 return NotFound();
             }
+
             return Ok(session);
         }
 
         // DELETE: api/PomodoroSession/5
         [HttpDelete("{id}")]
         public async Task<ActionResult<bool>> DeletePomodoroSession(int id)
+
         {
+            var userId = GetUserIdFromToken();
+            if (userId == null) return Unauthorized("Invalid User Token");
             var result = await _pomodoroSessionService.DeletePomodoroSession(id);
             if (!result)
             {
                 return NotFound();
             }
+
             return Ok(result);
         }
+
+        private string? GetUserIdFromToken()
+        {
+            return User.FindFirstValue(ClaimTypes.NameIdentifier);
+        }
+        }
     }
-}
